+++ mathics/builtin/drawing/image.py
@@ -2079,11 +2079,17 @@ class Image(Atom):
     def default_format(self, evaluation, form):
         return "-Image-"
 
-    def get_sort_key(self, pattern_sort=False):
+    # FIXME: return type should be a specific kind of Tuple, not a list.
+    def get_sort_key(self, pattern_sort=False) -> list:
         if pattern_sort:
+            # If pattern_sort=True, returns the sort key that matches to an Atom.
             return super(Image, self).get_sort_key(True)
         else:
-            return hash(self)
+            # If pattern is False, return a sort_key for the expression `Image[]`,
+            # but with a `2` instead of `1` in the 5th position,
+            # and adding two extra fields: the length in the 5th position,
+            # and a hash in the 6th place.
+            return [1, 3, SymbolImage, tuple(), 2, len(self.pixels), hash(self)]
 
     def sameQ(self, other) -> bool:
         """Mathics SameQ"""